# portfolio.node/src/docs/openapi.yml
openapi: 3.1.0
info:
  title: Portfolio Backend API
  description: |
    Backend API for the portfolio application following Domain Driven Design principles.
    - Frontend project: portfolio.next
    - Backend project: portfolio.node
    - Backend is supported by MongoDB, Qdrant and Ollama
  version: 1.5.0
...
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    DomainError:
      description: Domain-specific error
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                description: Domain-specific error message
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Authentication required"
    ForbiddenError:
      description: Admin access required
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Admin access required"
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Invalid request data"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
                example: "Resource not found"
  schemas:
...
    Project:
      type: object
      description: Project model representing a portfolio project
      required:
        - title
        - excerpt
        - description
      properties:
        id:
          type: string
          description: Unique identifier for the project
        title:
          type: string
          description: Project title
          minLength: 1
          maxLength: 200
        excerpt:
          type: string
          description: Short summary of the project
          minLength: 1
          maxLength: 500
        description:
          type: string
          description: Detailed project description in markdown format
          minLength: 1
        isDraft:
          type: boolean
          description: Whether the project is in draft status
          default: true
        publishAt:
          type: string
          format: date-time
          description: When to publish the project
        createdAt:
          type: string
          format: date-time
          description: When the project was created
        updatedAt:
          type: string
          format: date-time
          description: When the project was last updated
        tags:
          type: array
          description: List of technologies used in the project
          items:
            type: string
        links:
          type: object
          description: External links related to the project
          properties:
            github:
              type: string
              format: uri
            demo:
              type: string
              format: uri
        slug:
          type: string
          description: URL-friendly version of the title
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
...
    Page:
      type: object
      description: Static page model for custom content pages
      required:
        - title
        - content
        - slug
      properties:
        id:
          type: string
          description: Unique identifier for the page
        title:
          type: string
          description: Page title
          minLength: 1
          maxLength: 200
        content:
          type: string
          description: Page content in markdown format
          minLength: 1
        slug:
          type: string
          description: URL-friendly identifier for the page
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        isDraft:
          type: boolean
          description: Whether the page is in draft status
          default: true
        createdAt:
          type: string
          format: date-time
          description: When the page was created
        updatedAt:
          type: string
          format: date-time
          description: When the page was last updated
...
paths:
...
  /projects:
    get:
      tags: [Public]
      summary: Get all published projects
      description: Get a list of all published, non-draft projects
      responses:
        200:
          description: List of published projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        404:
          $ref: '#/components/responses/NotFoundError'
  /projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Public]
      summary: Get published project by ID
      description: Get a specific published project by its ID
      responses:
        200:
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          $ref: '#/components/responses/NotFoundError'
...
  /pages/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          description: URL-friendly identifier for the page
    get:
      tags: [Public]
      summary: Get published page
      description: Get a specific published page by its slug
      responses:
        200:
          description: Page content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        400:
          $ref: '#/components/responses/ValidationError'
        404:
          $ref: '#/components/responses/NotFoundError'
...
  /admin/projects:
    get:
      tags: [Projects]
      summary: Get all projects
      description: |
        Get all projects regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: List of all projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Projects]
      summary: Create new project
      description: |
        Create a new project entry.
        Defaults to draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        201:
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
...
  /admin/pages:
    get:
      tags: [Pages]
      summary: Get all pages
      description: |
        Get all pages regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: List of all pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Page'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Pages]
      summary: Create new page
      description: |
        Create a new static page.
        Defaults to draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Page'
      responses:
        201:
          description: Page created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
...
  /admin/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          description: Project identifier
    get:
      tags: [Projects]
      summary: Get project by ID
      description: |
        Get a specific project regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    put:
      tags: [Projects]
      summary: Update project
      description: |
        Update a project.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        200:
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    delete:
      tags: [Projects]
      summary: Delete project
      description: |
        Permanently delete a project.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        204:
          description: Project deleted successfully
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
  /admin/pages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          description: Page identifier
    get:
      tags: [Pages]
      summary: Get page by ID
      description: |
        Get a specific page regardless of draft status.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        200:
          description: Page details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    put:
      tags: [Pages]
      summary: Update page
      description: |
        Update a page.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Page'
      responses:
        200:
          description: Page updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
    delete:
      tags: [Pages]
      summary: Delete page
      description: |
        Permanently delete a page.
        Requires admin authentication.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        204:
          description: Page deleted successfully
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          description: Page not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/DomainError'
...